backend/requirements.txt
asgiref==3.7.2
Django==4.2.10
django-cors-headers==4.7.0
djangorestframework==3.14.0
sqlparse==0.5.3
cloudinary==1.36.0
django-cloudinary-storage==0.3.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.10
requests==2.32.3
dj-database-url==2.1.0
python-dotenv==1.0.0